Move common code to CompressionUtils
